code clean

+++ Entities/OrderEntity.cs
@@ -4,7 +4,7 @@
 public class OrderEntity : BaseEntity {
 	public string? Description { get; set; }
 	public OrderStatuses? Status { get; set; }
-	public decimal? Price { get; set; }
+	public double? Price { get; set; }
 	public DateTime? PayDateTime { get; set; }
 	public string? PayNumber { get; set; }
 	public DateTime? ReceivedDate { get; set; }
@@ -19,7 +19,7 @@ public class OrderReadDto {
 	public Guid? Id { get; set; }
 	public string? Description { get; set; }
 	public OrderStatuses? Status { get; set; }
-	public decimal? Price { get; set; }
+	public double? Price { get; set; }
 	public DateTime? PayDateTime { get; set; }
 	public string? PayNumber { get; set; }
 	public DateTime? ReceivedDate { get; set; }
